confmod-app/backend/pyproject.toml
[tool.poetry]
package-mode = false
name = "confmod-app-backend"
version = "0.1.0"
description = "Backend for the confmod-framework app"
authors = ["Maximilian Sudmann <maximilian.sudmann@rwth-aachen.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.111.0"
python-multipart = "^0.0.9"
libconfmod = {path = "../../confmod", develop = true }
pip = "^24.1.2"
sqlalchemy = "^2.0.31"
case-converter = "^1.1.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
